% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jiffy.R
\name{fetch}
\alias{fetch}
\title{Download a gif}
\usage{
fetch(x, what = "original", format = "gif", dest = NULL, quiet = FALSE)
}
\arguments{
\item{x}{A giphy object}

\item{what}{The type of image to download.  \code{original} will
always be available but giphy supplies \code{fixed_height},
\code{fixed_width} \code{downsized}, etc.  See
\code{names(x$images)} for possible options for a given image.
The default is to download the original (largest) gif.}

\item{format}{The format to download in.  Default is gif, but
\code{mp4} and \code{webp} are often available (for
\code{what="looping"} the only option is \code{mp4}.}

\item{dest}{Destination to download the file to.  Default is a
temporary file.}

\item{quiet}{Logical indicating if progress bars during download
should be suppressed.}
}
\value{
The filename of the downloaded file.  This can be fed into
  \code{\link{show_gif}}.
}
\description{
Download a gif
}
\author{
Rich FitzJohn
}

