% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giphy_api.R
\name{giphy_api_id}
\alias{giphy_api_id}
\title{Retrieve gif by id}
\usage{
giphy_api_id(ids, type = "gifs")
}
\arguments{
\item{ids}{A vector of ids}

\item{type}{Either "gifs" or "stickers" (or a partial string match
to either).  The default is to return gifs.}
}
\description{
Retrieve gif by id.  Less useful in general because you need to
know the id.  But this is useful for getting a specific gif.
}
\details{
Note that all the \code{giphy_api_*} functions do not actually
download gifs.
}

