% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giphy_api.R
\name{giphy_api_random}
\alias{giphy_api_random}
\title{Retrieve a random gif, by tag.}
\usage{
giphy_api_random(tag = NULL, rating = NULL, type = "gifs")
}
\arguments{
\item{tag}{The gif tag (much narrower set of allowed words than
\code{phrase} in \code{\link{giphy_api_search}} and
\code{\link{giphy_api_translate}}.}

\item{rating}{Limits gifs to a rating of y, g, pg, pg-13 or r
(this is a limit not a filter).}

\item{type}{Either "gifs" or "stickers" (or a partial string match
to either).  The default is to return gifs.}
}
\description{
Retrieve a random gif, by tag.
}
\details{
Note that all the \code{giphy_api_*} functions do not actually
download gifs.
}
\author{
Rich FitzJohn
}

