% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giphy_api.R
\name{giphy_api_search}
\alias{giphy_api_search}
\title{Search the giphy archives}
\usage{
giphy_api_search(phrase, n = NULL, offset = NULL, rating = NULL,
  type = "gifs")
}
\arguments{
\item{phrase}{Phrase to search for (required)}

\item{n}{Maximum number of images to return.}

\item{offset}{Starting point in the stream to search from (by
default, starts at the beginning of the stream, but specify 50
to start at the 50th gif).}

\item{rating}{Limits gifs to a rating of y, g, pg, pg-13 or r
(this is a limit not a filter).}

\item{type}{Either "gifs" or "stickers" (or a partial string match
to either).  The default is to return gifs.}
}
\value{
A list of results that is a little complicated at the
  present.
}
\description{
Search the giphy archives.
}
\details{
If \code{n} is omitted, giphy will return 25 images.

Note that all the \code{giphy_api_*} functions do not actually
download gifs.
}
\author{
Rich FitzJohn
}

