% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giphy_api.R
\name{giphy_api_translate}
\alias{giphy_api_translate}
\title{Translate words to gifs}
\usage{
giphy_api_translate(phrase, rating = NULL, type = "gifs")
}
\arguments{
\item{phrase}{Search word or phrase}

\item{rating}{Limits gifs to a rating of y, g, pg, pg-13 or r
(this is a limit not a filter).}

\item{type}{Either "gifs" or "stickers" (or a partial string match
to either).  The default is to return gifs.}
}
\value{
A list of results that is a little complicated at the moment
}
\description{
Intelligently translate phrases into gifs (in the way used by
Slack's \code{/giphy} add-in).  Giphy adds some "special sauce" to
make the results more relevant.
}
\details{
Note that all the \code{giphy_api_*} functions do not actually
download gifs.
}
\author{
Rich FitzJohn
}

