% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giphy_api.R
\name{giphy_api_trending}
\alias{giphy_api_trending}
\title{Retrieve trending gifs}
\usage{
giphy_api_trending(n = 25, rating = NULL, type = "gifs")
}
\arguments{
\item{n}{Maximum number of images to return.}

\item{rating}{Limits gifs to a rating of y, g, pg, pg-13 or r
(this is a limit not a filter).}

\item{type}{Either "gifs" or "stickers" (or a partial string match
to either).  The default is to return gifs.}
}
\description{
Retrieve trending gifs.
}
\details{
Note that all the \code{giphy_api_*} functions do not actually
download gifs.
}

