% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jiffy.R
\name{jiffy_random}
\alias{jiffy_random}
\alias{jiffy_translate}
\title{Download a random gif from giphy}
\usage{
jiffy_random(tag = NULL, type = "gifs", what = "original",
  format = "gif", dest = NULL, quiet = FALSE)

jiffy_translate(phrase, type = "gifs", what = "original", format = "gif",
  dest = NULL, quiet = FALSE)
}
\arguments{
\item{tag}{An optional tag to search on.  If omitted then you git
a random image pulled from the giphy's trending stream (I
believe).}

\item{type}{Either "gifs" or "stickers" (or a partial string match
to either).  The default is to return gifs.}

\item{what}{The type of image to download.  \code{original} will
always be available but giphy supplies \code{fixed_height},
\code{fixed_width} \code{downsized}, etc.  See
\code{names(x$images)} for possible options for a given image.
The default is to download the original (largest) gif.}

\item{format}{The format to download in.  Default is gif, but
\code{mp4} and \code{webp} are often available (for
\code{what="looping"} the only option is \code{mp4}.}

\item{dest}{Destination to download the file to.  Default is a
temporary file.}

\item{quiet}{Logical indicating if progress bars during download
should be suppressed.}

\item{phrase}{A phrase to search for}
}
\value{
A filename of the downloaded file.
}
\description{
Download a random gif from giphy.
}
\examples{
\dontrun{
# Download a random cat:
gif <- jiffy_random("cat")
show_gif(gif)

# MRW someone suggests not using stringsAsFactors=FALSE
gif <- jiffy_translate("hell no")
show_gif(gif)
}
}

