% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{show_gif}
\alias{show_gif}
\title{Display a gif}
\usage{
show_gif(filename, method = "browser")
}
\arguments{
\item{filename}{Filename of gif}

\item{method}{One of "browser", "shiny" or "gadget".  If you have
a recent version of shiny and are running in Rstudio, use method
"gadget" to open the gif in the viewer window.}
}
\description{
Display a gif using shiny
}

